src/cultural_insights.py
```python
from models.cultural_model import CulturalModel
from database.cultural_db import CulturalDB

class CulturalInsights:
    def __init__(self):
        self.cultural_model = CulturalModel()
        self.cultural_db = CulturalDB()

    def provideCulturalInsight(self, topic):
        # Fetch cultural insight from the database
        cultural_insight = self.cultural_db.get_cultural_insight(topic)

        if cultural_insight:
            # If cultural insight exists, return it
            return cultural_insight
        else:
            # If cultural insight does not exist, generate it using the cultural model
            generated_insight = self.cultural_model.generate_insight(topic)

            # Save the generated insight to the database for future use
            self.cultural_db.save_cultural_insight(topic, generated_insight)

            return generated_insight
```