src/tests/test_cultural_insights.py
```python
import unittest
from src.cultural_insights import provideCulturalInsight
from src.models.cultural_model import CulturalModel

class TestCulturalInsights(unittest.TestCase):

    def setUp(self):
        self.cultural_model = CulturalModel()

    def test_provide_cultural_insight(self):
        # Test with a known cultural insight
        insight = "Shabbat"
        result = provideCulturalInsight(insight, self.cultural_model)
        self.assertIsNotNone(result)
        self.assertIsInstance(result, str)

        # Test with an unknown cultural insight
        insight = "Unknown"
        result = provideCulturalInsight(insight, self.cultural_model)
        self.assertIsNone(result)

    def test_cultural_model(self):
        # Test the cultural model's get_insight method
        insight = "Kosher"
        result = self.cultural_model.get_insight(insight)
        self.assertIsNotNone(result)
        self.assertIsInstance(result, str)

        # Test with an unknown cultural insight
        insight = "Unknown"
        result = self.cultural_model.get_insight(insight)
        self.assertIsNone(result)

if __name__ == '__main__':
    unittest.main()
```